requirements.txt
appdirs==1.4.3
click==6.7
dj-database-url==0.4.1
Django==1.10.4
djangorestframework==3.6.2
gunicorn==19.6.0
packaging==16.8
psycopg2==2.6.2
pyparsing==2.2.0
python-dotenv==0.6.4
pytz==2017.2
requests==2.13.0
six==1.10.0
whitenoise==3.2
